Properly import favorites from backup
When importing favorites, there's a validation that we only import stuff that matches connections. However, on first import it is quite probable that favorites will be imported _before_ connections (it is just depending on the order in which the backup json file puts the preference file names).
Thus, andFHEM will now import non favorites first. Then we can be sure that connections will be available, so the check will succeed.¯